Split Sidebar to multiple components
