+++ client/src/components/LoggedIn.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Sidebar from './Sidebar';
+import Sidebar from './sidebar/Sidebar';
 
 export default class LoggedIn extends Component {
   render() {
