+++ client/src/components/sidebar/SidebarHeader.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default function SidebarHeader(props) {
+  return (
+    <header className="sidebar-header">
+      <h1 className="sidebar-header-title">
+        Cobase<span className="sidebar-header-title__light">PRO</span>
+      </h1>
+    </header>
+  );
+}
