+++ client/src/components/sidebar/SidebarSectionTitle.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const SidebarSectionTitle = ({children}) => (
+  <h3 className="sidebar-navigation-title">{children}</h3>
+);
+
+export default SidebarSectionTitle;
